wk_geovi/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
redis==3.5.3
requests==2.25.1
rq==1.8.0
urllib3==1.26.4
Werkzeug==1.0.1
requests

sentinelsat
rasterio
shapely
python-dotenv
numpy
pyproj
geopandas

#requests==2.31.0
requests-oauthlib==1.3.1
Pillow==10.0.0
#python-dotenv==1.0.0

